Refactor: workspacemember 패키지를 workspace 패키지 밖으로 이동
- workspacemember 패키지를 workspace, member 패키지와 같은 레벨에 위치
