--- src/main/java/com/uranus/taskmanager/api/workspace/workspacemember/WorkspaceRole.java
@@ -1,7 +1,0 @@
-package com.uranus.taskmanager.api.workspace.workspacemember;
-
-public enum WorkspaceRole {
-	ADMIN,
-	USER,
-	READER
-}
