+++ src/main/java/com/uranus/taskmanager/api/workspacemember/WorkspaceRole.java
@@ -1,0 +1,7 @@
+package com.uranus.taskmanager.api.workspacemember;
+
+public enum WorkspaceRole {
+	ADMIN,
+	USER,
+	READER
+}
