+++ src/main/java/com/uranus/taskmanager/api/workspacemember/repository/WorkspaceMemberRepository.java
@@ -1,4 +1,4 @@
-package com.uranus.taskmanager.api.workspace.workspacemember.repository;
+package com.uranus.taskmanager.api.workspacemember.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.uranus.taskmanager.api.member.domain.Member;
 import com.uranus.taskmanager.api.workspace.domain.Workspace;
-import com.uranus.taskmanager.api.workspace.workspacemember.domain.WorkspaceMember;
+import com.uranus.taskmanager.api.workspacemember.domain.WorkspaceMember;
 
 public interface WorkspaceMemberRepository extends JpaRepository<WorkspaceMember, Long> {
 	List<WorkspaceMember> findByWorkspaceId(Long workspaceId);
